requirements.txt

pyserial==3.5
configobj==5.0.8
paho-mqtt==1.6.1